chore: Consolidate auth system and clean up providers
- Rename SimpleAuthProvider.tsx to AuthProvider.tsx (the main one)
- Delete old AuthProvider.tsx and AuthProvider-v2.tsx
- Update all imports to use the consolidated AuthProvider
- Consolidate auth utilities into single auth-utils.ts file
- Delete redundant auth-recovery.ts file
- Update test setup to use new AuthProvider

This eliminates confusion and provides a single, clean auth system
with proper profile fetching and error handling.
