--- src/contexts/AuthProvider-v2.tsx
@@ -1,192 +1,0 @@
-import React, { createContext, useContext, useEffect, useState } from 'react';
-import { User, Session } from '@supabase/supabase-js';
-import { supabase, type Profile } from '@/lib/supabase';
-
-// Auth context type
-interface AuthContextType {
-  user: User | null;
-  profile: Profile | null;
-  session: Session | null;
-  loading: boolean;
-  refreshProfile: () => Promise<void>;
-  signOut: () => Promise<void>;
-}
-
-// Create context
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
-
-// Custom hook to use auth context
-export function useAuth(): AuthContextType {
-  const context = useContext(AuthContext);
-  if (context === undefined) {
-    throw new Error('useAuth must be used within an AuthProvider');
-  }
-  return context;
-}
-
-// AuthProvider component
-export function AuthProvider({ children }: { children: React.ReactNode }) {
-  const [user, setUser] = useState<User | null>(null);
-  const [profile, setProfile] = useState<Profile | null>(null);
-  const [session, setSession] = useState<Session | null>(null);
-  const [loading, setLoading] = useState(true);
-
-  // Function to fetch user profile
-  const fetchProfile = async (userId: string): Promise<Profile | null> => {
-    try {
-      const { data, error } = await supabase
-        .from('profiles')
-        .select('id, username, full_name, avatar_url, created_at, updated_at')
-        .eq('id', userId)
-        .single();
-
-      if (error) {
-        // If the profiles table doesn't exist yet (migrations not run), fail gracefully
-        if (error.code === '42P01') {
-          console.warn(
-            'Profiles table does not exist yet. Please run database migrations.'
-          );
-          return null;
-        }
-        console.error('Error fetching profile:', error);
-        return null;
-      }
-
-      return data;
-    } catch (error) {
-      console.error('Error in fetchProfile:', error);
-      return null;
-    }
-  };
-
-  // Function to refresh profile data
-  const refreshProfile = async (): Promise<void> => {
-    if (!user) {
-      setProfile(null);
-      return;
-    }
-
-    const profileData = await fetchProfile(user.id);
-    setProfile(profileData);
-  };
-
-  // Function to sign out
-  const signOut = async (): Promise<void> => {
-    try {
-      const { error } = await supabase.auth.signOut();
-      if (error) {
-        console.error('Sign out error:', error);
-        throw error;
-      }
-    } catch (error) {
-      console.error('Sign out error:', error);
-      throw error;
-    }
-  };
-
-  // Initialize auth state and set up auth listener
-  useEffect(() => {
-    let mounted = true;
-
-    // Get initial session
-    const initializeAuth = async (): Promise<void> => {
-      try {
-        const {
-          data: { session },
-          error,
-        } = await supabase.auth.getSession();
-
-        if (error) {
-          console.error('Error getting session:', error);
-          // If there's a session error, clear any stale tokens
-          if (
-            error.message?.includes('Invalid Refresh Token') ||
-            error.message?.includes('Refresh Token Not Found')
-          ) {
-            console.log('🔄 Clearing stale tokens due to refresh token error');
-            await supabase.auth.signOut();
-          }
-          if (mounted) {
-            setLoading(false);
-          }
-          return;
-        }
-
-        if (mounted) {
-          setSession(session);
-          setUser(session?.user ?? null);
-
-          // Fetch profile if user exists
-          if (session?.user) {
-            const profileData = await fetchProfile(session.user.id);
-            if (mounted) {
-              setProfile(profileData);
-            }
-          }
-
-          setLoading(false);
-        }
-      } catch (error) {
-        console.error('Error initializing auth:', error);
-        if (mounted) {
-          setLoading(false);
-        }
-      }
-    };
-
-    initializeAuth();
-
-    // Set up auth state change listener
-    const {
-      data: { subscription },
-    } = supabase.auth.onAuthStateChange(async (event, session) => {
-      if (!mounted) return;
-
-      // Handle token refresh errors
-      if (event === 'TOKEN_REFRESHED' && !session) {
-        console.warn('Token refresh failed, clearing auth state');
-        setSession(null);
-        setUser(null);
-        setProfile(null);
-        setLoading(false);
-        return;
-      }
-
-      setSession(session);
-      setUser(session?.user ?? null);
-
-      // Handle profile data based on auth event
-      if (session?.user) {
-        // User signed in - fetch their profile
-        const profileData = await fetchProfile(session.user.id);
-        if (mounted) {
-          setProfile(profileData);
-        }
-      } else {
-        // User signed out - clear profile
-        setProfile(null);
-      }
-
-      // Auth is no longer loading after first state change
-      setLoading(false);
-    });
-
-    // Cleanup function
-    return () => {
-      mounted = false;
-      subscription.unsubscribe();
-    };
-  }, []);
-
-  // Context value
-  const value: AuthContextType = {
-    user,
-    profile,
-    session,
-    loading,
-    refreshProfile,
-    signOut,
-  };
-
-  return React.createElement(AuthContext.Provider, { value }, children);
-}
